🥅  Request validation (#8)
This PR refactors the request validation to explicitly enumerate the
errors and messages that TGIS currently implements. It implements most
of the missing validation, excluding `PromptPrefix` and `Tokenizer` at
the moment, since soft prompts are not supported yet and I'm not 100%
sure how exactly to wire up the tokenizer one without running this and
figuring out how to trigger it.

I'd like to add some basic unit tests here and get a GHA workflow going
to run them. (The big blocker to that in general here is the need for
GPUs, but I think we can test this logic without needing to load a real
model)

---------

Signed-off-by: Joe Runde <Joseph.Runde@ibm.com>