Add a sequence id to the logits processor signature
This allows the implementation of stateful logit processors.
By passing the sequence ID, the decision on how to manage the state
is the responsibility of the author of the processor.
Without a sequence ID we would have to defensively create a copy
of the processor for each sequence just to safeguard against
accidental state sharing.

Signed-off-by: Max de Bayser <mbayser@br.ibm.com>