Add github workflow to run Spyre tests (#59)
This PR enables the Spyre tests to run as a Github action. 

I realized that the model we were using for the tests `llama-194m` is
not available on HF hub, but if we want to run the tests externally we
need to use some model that is available. I've replaced it with this
one: https://huggingface.co/JackFram/llama-160m

Note I haven't actually changed the model name in the tests, I just
"hacked" it for now using a soft link in the docker container. This is
because there is some ongoing work to introduce environment variables to
control the tests and I don't want to complicate things.

For this model I see some quite weird behaviour where the tokens
produced by vLLM and HF Transformers are identical but the decode text
is slightly different (they are the same up to a leading space). I don't
think this difference is related to Spyre so I've just changed the test
to compare token ids instead.

---------

Signed-off-by: Thomas Parnell <tpa@zurich.ibm.com>