Refactor code according to upstream changes  (#62)
This PR reworks our code according to some important upstream changes.
In particular, there is no longer any need to have a separate
`SpyreExecutor` and `MultiprocessingSpyreExecutor`. Upstream has added
generic classes for this that work across different platforms. Acutally,
it simplifies our code quite a lot.

The model runner classes now inherit from `ModelRunnerBase` and we need
to define a `ModelInputForSpyre` class accordingly.

This is current passing all CPU tests, but needs to be tested on Spyre
and needs careful review since it quite a big change.


**Note:** the target for this PR is a branch `upstream-2025-01-17`
containing upstream changes merged into our current branch. I've done it
like this so it is easier to review the changes. If this PR is approved,
we can then merge the changes into `upstream-2025-01-17` and then merge
that one into main.

---------

Signed-off-by: Thomas Parnell <tpa@zurich.ibm.com>
Signed-off-by: Yannick Schnider <Yannick.Schnider1@ibm.com>
Co-authored-by: Yannick Schnider <Yannick.Schnider1@ibm.com>