Initial support for span semantics (#88)
## Purpose

An initial implementation of span semantics in vLLM. 

Please note that this has a known bug dealing with concurrent sequences
that re-use the same span in different locations. We are working on a
solution for this, but in the meantime accuracy may be negatively
affected.

## Test Plan

n/a

## Test Result

n/a

---
<details>
<summary> Essential Elements of an Effective PR Description Checklist
</summary>

- [x] The purpose of the PR, such as "Fix some issue (link existing
issues this PR will resolve)".
- [ ] The test plan, such as providing test command.
- [ ] The test results, such as pasting the results comparison before
and after, or e2e results
- [ ] (Optional) The necessary documentation update, such as updating
`supported_models.md` and `examples` for a new model.
- [ ] (Optional) Release notes update. If your change is user facing,
please update the release notes draft in the [Google
Doc](https://docs.google.com/document/d/1YyVqrgX4gHTtrstbq8oWUImOyPCKSGnJ7xtTpmXzlRs/edit?tab=t.0).
</details>

---------

Signed-off-by: Thomas Parnell <tpa@zurich.ibm.com>
Signed-off-by: Nathan Ordonez <nathanaxcan@gmail.com>
Co-authored-by: Nathan Ordonez <Nathan.Ordonez@ibm.com>
Co-authored-by: Nathan Ordonez <nathanaxcan@gmail.com>